Remove singleton logic cones from set cover formulation
We no longer seek to cover *all* vertices of a graph. Instead we want
a partial cover of maximum size. This helps reduce the size of the ILP
formulation (without changing its runtime)

We also augment the Cone structure to keep track of its primary
inputs (the cut that resulted in the cone).