Fix excessive NOPs in scheduler
There are excessive NOPs classified as "stall for recv" in the core
that has finished its computation. The reason is that this core has
finished, but the other cores' state is still advancing, so the scheduler
keeps inserting NOPs in the finished core (which is correct, though not
needed due to the epilogue counters that the machine code generator later
sets), but the issue is that these NOPs get classified as "wait for RECV",
whereas in reality they are "sleep" NOPs (which the scheduler does not see).